fix: better incapsulation
